requirements.txt
# =============================================================================
# PRIMUS BACKEND - COMPREHENSIVE REQUIREMENTS FOR UBUNTU SERVER DEPLOYMENT
# =============================================================================

# Core FastAPI Framework & ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
starlette==0.27.0

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
databases[postgresql]==0.8.0
alembic==1.12.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]==2.5.0

# HTTP Client & API Integration
httpx==0.25.2
requests==2.31.0
authlib==1.2.1

# WebSocket Support
websockets==12.0
websocket-client==1.6.4

# Firebase & Google Cloud
firebase-admin==6.2.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Payment Processing
stripe==7.8.0
razorpay==1.4.2

# Email & Communication
email-validator==2.1.0

# Image Processing & Screenshots
pillow==10.1.0
mss==9.0.1

# PDF Generation
reportlab==4.0.7

# Redis (Optional - for OTP caching)
redis==5.0.1

# Date/Time Utilities
python-dateutil==2.8.2

# JSON Web Tokens
PyJWT==2.8.0

# CSV Processing
csv-writer==1.0.0

# File Upload & Processing
python-magic==0.4.27

# =============================================================================
# PRODUCTION & DEPLOYMENT PACKAGES FOR UBUNTU SERVER
# =============================================================================

# Production WSGI/ASGI Servers
gevent==23.9.1
eventlet==0.33.3

# Process Management
supervisor==4.2.5

# System Monitoring & Logging
psutil==5.9.6
structlog==23.2.0
rich==13.7.0

# Database Connection Pooling
SQLAlchemy-Utils==0.41.1
psycopg2==2.9.9

# SSL/TLS Support
pyOpenSSL==23.3.0
cryptography==41.0.8

# Environment & System
python-decouple==3.8
click==8.1.7

# Testing (for CI/CD)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Performance & Caching
aioredis==2.0.1
hiredis==2.2.3

# Background Task Processing
celery==5.3.4
kombu==5.3.4

# File System Utilities
watchdog==3.0.0

# Network & Security
certifi==2023.11.17
urllib3==2.1.0

# Data Serialization
orjson==3.9.10
msgpack==1.0.7

# Timezone Support
pytz==2023.3.post1

# =============================================================================
# UBUNTU-SPECIFIC SYSTEM DEPENDENCIES (install via apt)
# =============================================================================
# Run these commands on Ubuntu server before pip install:
#
# sudo apt update
# sudo apt install -y python3-dev python3-pip python3-venv
# sudo apt install -y postgresql postgresql-contrib libpq-dev
# sudo apt install -y redis-server
# sudo apt install -y nginx
# sudo apt install -y supervisor
# sudo apt install -y build-essential
# sudo apt install -y libssl-dev libffi-dev
# sudo apt install -y libjpeg-dev libpng-dev
# sudo apt install -y pkg-config
# sudo apt install -y git curl wget
# sudo apt install -y certbot python3-certbot-nginx
#
# =============================================================================

# Development Tools (remove in production)
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.1
# pre-commit==3.6.0

# =============================================================================
# INSTALLATION NOTES FOR UBUNTU SERVER:
# =============================================================================
# 
# 1. Create virtual environment:
#    python3 -m venv venv
#    source venv/bin/activate
#
# 2. Upgrade pip:
#    pip install --upgrade pip setuptools wheel
#
# 3. Install requirements:
#    pip install -r requirements.txt
#
# 4. Set up PostgreSQL database:
#    sudo -u postgres createuser primus
#    sudo -u postgres createdb primus_db -O primus
#    sudo -u postgres psql -c "ALTER USER primus PASSWORD 'your_password';"
#
# 5. Configure environment variables in .env file
#
# 6. Run database migrations:
#    alembic upgrade head
#
# 7. Start with Gunicorn (production):
#    gunicorn -w 4 -k uvicorn.workers.UvicornWorker main:app --bind 0.0.0.0:8000
#
# 8. Configure Nginx as reverse proxy
#
# 9. Set up SSL with Let's Encrypt:
#    sudo certbot --nginx -d yourdomain.com
#
# =============================================================================